pyproject.toml
[project]
name = "notion-mcp-server"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "langchain-mcp-adapters>=0.2.2",
    "python-dotenv>=1.2.2",
]
